Replace list markers in goal list with emojis
If you send a multi-line list as an embed field to Discord,
it will be rendered with an extra line between the list items.
To avoid this issue, this commit replaces list markers with emojis.
Different emojis are shown depending on if the goal is complete,
in progress, or not started.